CSSE-13: add consistency with azure hybrid docs
Updating the azure hybrid docs so that the CLI based deployment method
uses the same resource names as the terraform based deployment. These
are the new values in the docs. Specifically we're removing any mention
of "demo" from the strings so that users can use this for a prod like
deployment.

GitOrigin-RevId: 01b24b09780afad0e78fefa2a328415e56172f19